+++ changelog.md
@@ -1,5 +1,10 @@
 # Rest Recovery Changelog
 
+## Version 1.3.8
+- Added UI to prompt rests for individual players, button can be found in the bottom left corner, which can be hidden in the module settings
+- Added `game.restrecovery.renderPromptRestInterface`, which can render the UI mentioned above through code
+- Fixed bug where opening and closing the settings UI would cause it to fail to save and load module profiles
+
 ## Version 1.3.7
 - Moved short rest pact slot recovery setting to the short rest section